requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
Brotli==1.1.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
dj-database-url==2.2.0
Django==5.1.1
django-browser-reload==1.15.0
django-phonenumber-field==8.0.0
django-tailwind==3.8.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysql-connector-python==9.0.0
mysqlclient==2.2.4
packaging==24.1
phonenumberslite==8.13.45
psycopg2-binary==2.9.10
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.32.0
whitenoise==6.7.0
